LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/setrequestheader-open-setrequestheader-expected.txt


FAIL XMLHttpRequest: setRequestHeader() and open() assert_equals: Set headers record should have been cleared by open. expected "X-appended-to-this: True\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"/Volumes/Data/EWS/WebKit/LayoutTests/imported/w3c/web-platform-tests/tools/wptserve/wptserve/handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"/Volumes/Data/EWS/WebKit/LayoutTests/imported/w3c/web-platform-tests/XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
